Methods - Internal

Index

EnergyModelsFlex.check_limits_default

EnergyModelsFlex.check_period_ts

Check methods

EnergyModelsFlex.check_period_ts — Function

check_period_ts(ts::RepresentativePeriods, n::PeriodDemandSink, msg::String)
check_period_ts(ts::OperationalScenarios, n::PeriodDemandSink, msg::String)
check_period_ts(ts::SimpleTimes, n::PeriodDemandSink, msg::String)

Function for checking that the timestructure is valid in combination with the chosen period structure in a [PeriodDemandSink(@ref).

EnergyModelsFlex.check_limits_default — Function

check_limits_default(n::Union{LimitedFlexibleInput, Combustion})

This function checks that the limits of a LimitedFlexibleInput or Combustion node are valid.