Methods - Accessing fields

Index

EnergyModelsFlex.activation_consumption

EnergyModelsFlex.number_of_periods

EnergyModelsFlex.period_demand

EnergyModelsFlex.period_index

EnergyModelsFlex.period_length

PeriodDemandSink types

EnergyModelsFlex.period_demand — Function

period_demand(n::AbstractPeriodDemandSink)
period_demand(n::AbstractPeriodDemandSink, i::Int)

Returns the period demand of AbstractPeriodDemandSink n as Array or in demand period i.

EnergyModelsFlex.period_length — Function

period_length(n::AbstractPeriodDemandSink)

Returns the length of the demand period of AbstractPeriodDemandSink n.

EnergyModelsFlex.period_index — Function

period_index(n::AbstractPeriodDemandSink, t)

Returns the index of the period (e.g., day) that a operational period t belongs to.

EnergyModelsFlex.number_of_periods — Function

number_of_periods(n::AbstractPeriodDemandSink)
number_of_periods(n::AbstractPeriodDemandSink, 𝒯::TimeStructure)

Returns the number of periods for a PeriodDemandSink n. If a TimeStructure is provided it calculates it based on the chosen time structure.

ActivationCostNode types

EnergyModelsFlex.activation_consumption — Function

activation_consumption(n::ActivationCostNode)
activation_consumption(n::ActivationCostNode, p::Resource)

Returns the demand during activation of ActivationCostNode n as dictionary or for input Resource p. If p is not included in the dictionary, it returns a value of 0.