Types

Index

EnergyModelsFlex.AbstractMultipleInputSink

EnergyModelsFlex.AbstractMultipleInputSinkStrat

EnergyModelsFlex.AbstractPeriodDemandSink

EnergyModelsFlex.UnitCommitmentNode

Nodal supertypes

EnergyModelsFlex.UnitCommitmentNode — Type

UnitCommitmentNode{} <: EMB.NetworkNode

Abstract type for unit commitment nodes.

EnergyModelsFlex.AbstractPeriodDemandSink — Type

abstract type AbstractPeriodDemandSink <: EMB.Sink

Supertypes for period demand sinks in which the demand must be satisifed within a given period.

EnergyModelsFlex.AbstractMultipleInputSink — Type

abstract type AbstractMultipleInputSink <: Sink

Abstract supertype for Sink nodes in which the demand can be satisfied by multiple resources.

EnergyModelsFlex.AbstractMultipleInputSinkStrat — Type

abstract type AbstractMultipleInputSinkStrat <: AbstractMultipleInputSink

Abstract supertype for AbstractMultipleInputSink nodes in which the ratio between the different resources must be constant within a strategic period.